requirements.txt
streamlit==1.36.0
streamlit-chat==0.1.1
# --- Wersje LangChain obniżone w celu zapewnienia kompatybilności ---
langchain==0.1.20
langchain-openai==0.1.0
# Usunięto langchain-text-splitters - w tej wersji jest częścią langchain
# -----------------------------------------------------------------
openai==1.35.7
beautifulsoup4==4.12.3
pypdf==4.2.0
lxml==5.2.2
html5lib==1.1
xhtml2pdf==0.2.15
pydantic==2.7.1
# --- Wersje bibliotek Google obniżone w celu rozwiązania błędu 504 ---
google-generativeai==0.5.4
langchain-google-genai==1.0.3
# -------------------------------------------------------------------
chromadb==0.5.3
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab
PyMuPDF
Pillow<11 